fix(test): use io.quarkus.test.InjectMock (package moved in Quarkus 3)
@InjectMock moved from io.quarkus.test.junit.mockito.InjectMock to
io.quarkus.test.InjectMock in Quarkus 3.x; the old package no longer
exists in Quarkus 3.29, so build-jvm/build-native failed to compile the
*ResourceErrorTest / *ResourceTest classes with "cannot find symbol:
class InjectMock". This surfaced only now because Java CI had been
disabled while the platform was bumped to 3.29.

Update the import in all 10 affected test classes. Verified locally:
jaxrs-rabbitmq-quarkus test-compile now succeeds.
