+++ application/jaxrs-activemq-quarkus/src/test/java/webapp/tier/resource/ActivemqResourceErrorTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import io.quarkus.artemis.test.ArtemisTestResource;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.mockito.InjectMock;
+import io.quarkus.test.InjectMock;
 import webapp.tier.service.ActiveMqService;
 
 @QuarkusTest
