+++ application/jaxrs-memcached-quarkus/src/test/java/webapp/tier/resource/MemcachedResourceErrorTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
 
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.mockito.InjectMock;
+import io.quarkus.test.InjectMock;
 import webapp.tier.service.MemcachedService;
 
 @QuarkusTest
