+++ application/jaxrs-mongodb-quarkus/src/test/java/webapp/tier/resource/MongodbResourceTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
 
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.mockito.InjectMock;
+import io.quarkus.test.InjectMock;
 import webapp.tier.bean.MsgBean;
 import webapp.tier.service.MongodbService;
 import webapp.tier.util.CreateId;
