+++ application/jaxrs-redis-quarkus/src/test/java/webapp/tier/resource/RedisResourceErrorTest.java
@@ -9,7 +9,7 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.mockito.InjectMock;
+import io.quarkus.test.InjectMock;
 import io.restassured.http.ContentType;
 import webapp.tier.service.RedisService;
