+++ application/randompublish-quarkus/src/test/java/webapp/tier/resource/RandomResourceErrorTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
 
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.mockito.InjectMock;
+import io.quarkus.test.InjectMock;
 import webapp.tier.service.RandomService;
 
 @QuarkusTest
